requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
twilio==9.0.0
anthropic==0.39.0
python-dotenv==1.0.0
